requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
openai==1.54.4
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyTelegramBotAPI==4.24.0
python-dotenv
requests==2.32.3
sniffio==1.3.1
SpeechRecognition==3.11.0
telebot==0.0.5
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
pytz==2024.2
kafka-python==2.0.2